+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### [0.14.2]
+- return keyword if no items were selected
+### [0.14.1]
+- add a new property `select-on-blur`
+### [0.14.0]
+- #####Upgraded to Angular 4
+### [0.13.4]
+- now the first item on the suggestion dropdown will not be auto-selected until we press the arrow down key
 ### [0.13.0]
 - renamed module to @ngui/auto-complete
 ### [0.12.0]
@@ -7,14 +15,14 @@
 ### [0.10.10]
 - fixed `accept-user-input`, fixed tests
 ### [0.10.9]
-- bug fox on list formatter not applying and dropdown positioning
+- bug fix on list formatter not applying and dropdown positioning
 - fixed (undefined) display when source is list of number or boolean
 ### [0.10.8]
 - fixed (undefined) display when source is list of number or boolean
 ### [0.10.7]
 - introduced `select-valule-of`
 - removed `value-property-name`
-- list-formatter now accepts sttirng e.g. `(id) value`
+- list-formatter now accepts string e.g. `(id) value`
 ### [0.10.6]
 - testing with initial focus
 ### [0.10.5]